revert: undo widgets container split into separate component files
Reverts the WidgetCard / PrintReportButton / WidgetsDeckButton
extraction from f6bc814d. The mobile-legend top-20 spacing fix and
the About panel font-sans typo fix from that commit are retained
in their respective files.

Reason: in CI, the split caused a Turbopack chunk-evaluation TDZ
(`Cannot access 'it' before initialization`) that surfaced as a
page console.error on every test and timed the suite out. The
inline rendering used in widgets/index.tsx before f6bc814d is
known-good in CI.
